+++ EOS/cryptojinian/config.hpp
@@ -28,13 +28,13 @@ constexpr int64_t MINING_COEF = 10 ;
 
 constexpr uint64_t PRICE_SCALE = 100000000;
 
-constexpr uint8_t MINING_TIMES = 16 ;
+constexpr uint32_t MINING_TIMES = 16 ;
 
 constexpr double_t DIVIDEND_COEF{0.65} ;    // 分红调成65%
 constexpr double_t BUYBACK_COEF{0.2} ;      // 回购调成20%
 constexpr double_t COLLECTION_COEF{0.15} ;  // 集齐整版纪念币奖励调成15%
 
-constexpr uint8_t FROZEN_DAYS = 30 ;
+constexpr uint32_t FROZEN_DAYS = 30 ;
 
 const asset cost_table( const uint64_t &n ) {
     if ( n > 408120 ) return asset( string_to_price("1.0000"), EOS_SYMBOL ) * 2 ;
@@ -85,7 +85,7 @@ const vector<vector<uint64_t>> _coinvalues = { // [][+1]
         {500,500,1000,2000,5000}    //tr 21
 };
 
-const vector<uint8_t> collection_combination_parameters( const uint8_t &type ) {
+const vector<uint32_t> collection_combination_parameters( const uint32_t &type ) {
     switch( type ) {
         case 22 :
             return {0, 2, 10};
@@ -106,7 +106,7 @@ const vector<uint8_t> collection_combination_parameters( const uint8_t &type ) {
     }
 }
 
-const asset bouns_table( const uint8_t &type ) {
+const asset bouns_table( const uint32_t &type ) {
     switch( type ) {
         case 0 :
             return asset( string_to_price("10000.0000"), TOKEN_SYMBOL ) ;
