+++ EOS/cryptojinian/cryptojinian.cpp
@@ -12,14 +12,14 @@ bool cryptojinian::cd_check( const uint64_t &id )
         
 }
 
-bool cryptojinian::cd_check( const name &owner, const uint8_t &type )
+bool cryptojinian::cd_check( const name &owner, const uint32_t &type )
 {
    singleton_collcd_t collcd(_self, owner.value);
    auto itr = collcd.get_or_create(_self, st_collection_cd { .time_limit = vector<uint32_t> (22 + 6 + 1, now()-1) } );
    return now() > itr.time_limit[type];
 }
 
-void cryptojinian::update_frozen_time_limit( const name &owner, const uint8_t &type, const uint64_t &quantity, const uint32_t &frozen_days )
+void cryptojinian::update_frozen_time_limit( const name &owner, const uint32_t &type, const uint64_t &quantity, const uint32_t &frozen_days )
 {
     frozencoins_t frozencoins(_self, _self.value);
     auto itr_p = _players.require_find(owner.value, "Player not found.") ;
@@ -48,7 +48,7 @@ void cryptojinian::update_frozen_time_limit( const name &owner, const uint8_t &t
    }
 }
 
-void cryptojinian::update_frozen_time_limit( const name &owner, const uint8_t &type, const uint32_t &frozen_days )
+void cryptojinian::update_frozen_time_limit( const name &owner, const uint32_t &type, const uint32_t &frozen_days )
 {
    singleton_collcd_t collcd(_self, owner.value);
    auto itr = collcd.get_or_create(_self, st_collection_cd { .time_limit = vector<uint32_t> (22 + 6 + 1, now()-1) } );
